django/requirements.txt
django==3.0.13
django-gcloud-connectors==0.3.2
django-htmlmin==0.11.0
google-cloud-logging==2.3.1
markdown==3.3.4
